removed intValue and longValue tests from PrimitiveTests and added to LazilyParsedNumberTests and refactored tests so that intValue acts like longValue and fails with BigInteger and passes with BigDecimal

+++ src/main/java/com/google/gson/internal/LazilyParsedNumber.java
@@ -17,6 +17,7 @@ package com.google.gson.internal;
 
 import java.io.ObjectStreamException;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * This class holds a number value that is lazily converted to a specific number type
@@ -39,7 +40,7 @@ public final class LazilyParsedNumber extends Number {
       try {
         return (int) Long.parseLong(value);
       } catch (NumberFormatException nfe) {
-        return new BigDecimal(value).intValue();
+        return new BigInteger(value).intValue();
       }
     }
   }
@@ -49,7 +50,7 @@ public final class LazilyParsedNumber extends Number {
     try {
       return Long.parseLong(value);
     } catch (NumberFormatException e) {
-      return new BigDecimal(value).longValue();
+      return new BigInteger(value).longValue();
     }
   }
